rescale:updated interpolation function

+++ rescale/rescale.h
@@ -7,7 +7,7 @@
 void interpolate(t_bw_img *input,
 		 t_bw_img *output,
 		 uint mask_len,
-		 float mask[][mask_len],
+		 double mask[][mask_len],
 		 double factor);
